all authentication done and debugging is done

+++ lib/models/product/cart_model.dart
@@ -7,7 +7,7 @@ class CartModel {
   });
 
   String? id;
-  double? qty;
+  int? qty;
   String? mrp;
   String? discountedPrice;
